pyproject.toml
[tool.poetry]
name = "helpfuldiscordbot"
version = "0.1.0"
description = ""
authors = ["Jackfreeman1 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.17.0"
boto3 = "^1.17.53"
discord = "^1.0.1"
flask = "^1.1.2"
finnhub-python = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
